requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
SQLAlchemy==2.0.23
python-dateutil==2.8.2
Flask-Login==0.6.2
python-telegram-bot==20.7
requests==2.31.0
APScheduler==3.10.4
